+++ src/main/java/acceso_datos/persistencia_bd/IControladorPBDPerfiles.java
@@ -1,0 +1,11 @@
+package acceso_datos.persistencia_bd;
+
+import java.math.BigInteger;
+import java.sql.SQLException;
+
+public interface IControladorPBDPerfiles {
+
+    public void eliminarPerfil(BigInteger idUsuario) throws SQLException;
+
+    public void modificarPerfil(BigInteger idUsuario, String valorAModificar, String modificacion) throws SQLException;
+}
